refactor: 단일 책임을 가지며, 메인 페이지는 레이아웃과 데이터 조합만 담당
